+++ jenkins/jenkins_jobs.groovy
@@ -1,3 +1,0 @@
-jobFactory.libraryJobs({
-requiredJDK "JAVA_11"
-})