+++ src/main/kotlin/ru/yoomoney/tech/grafana/dsl/panels/alerting/MaxCondition.kt
@@ -1,0 +1,25 @@
+package ru.yoomoney.tech.grafana.dsl.panels.alerting
+
+import org.json.JSONObject
+import ru.yoomoney.tech.grafana.dsl.json.set
+
+/**
+ * Generator for max alert condition.
+ * @author abramovgerman
+ * @since 02.12.2020
+ */
+class MaxCondition(private val condition: AlertingCondition) : AlertingCondition {
+
+    override fun toJson(): JSONObject {
+        val json = condition.toJson()
+
+        json["reducer"] = AlertEvaluator("max").toJson()
+
+        val operator = JSONObject()
+        operator["type"] = "and"
+
+        json["operator"] = operator
+
+        return json
+    }
+}