+++ src/main/kotlin/ru/yoomoney/tech/grafana/dsl/panels/stat/GraphMode.kt
@@ -1,0 +1,18 @@
+package ru.yoomoney.tech.grafana.dsl.panels.stat
+
+/**
+ * Graph display mode
+ * @author Aleksey Matveev
+ * @since 03.11.2020
+ */
+enum class GraphMode(val value: String) {
+    /**
+     * Hides the graph and only shows the value
+     */
+    NONE("none"),
+
+    /**
+     * Shows the area graph below the value. This requires that your query returns a time column
+     */
+    AREA("area");
+}