+++ src/test/kotlin/ru/yoomoney/tech/grafana/dsl/metrics/functions/RemoveAbovePercentileTest.kt
@@ -1,0 +1,20 @@
+package ru.yoomoney.tech.grafana.dsl.metrics.functions
+
+import org.amshove.kluent.shouldEqual
+import org.testng.annotations.Test
+
+/**
+ * Test for [ru.yoomoney.tech.grafana.dsl.metrics.functions.RemoveAbovePercentile]
+ * @author abramovgerman
+ * @since 02.12.2020
+ */
+class RemoveAbovePercentileTest {
+    @Test
+    fun `should create metric with removeAbovePercentile() function`() {
+        // given
+        val metric = "*.*.count".removeAbovePercentile(90)
+
+        // then
+        metric.asString() shouldEqual "removeAbovePercentile(*.*.count, 90)"
+    }
+}